Stop juggernaut feature from mysteriously vanishing
If a player were to quit and disappear without being eaten as juggernaut, then the game would falsely believe a juggernaut is still in play and that it should not create another juggernaut.  This lead to a condition where no player could become a juggernaut anymore.

This bug was based on the assumption that the only way a player would die is if they got eaten.  This did not take into account players quitting the game session.  This fix should cover any condition where a player leaves (because of being killed, quitting --  or any other future mechanics, thereby avoiding regressions).